# openapi.json
{"openapi":"3.1.0","info":{"title":"DY Crane Safety Management API","description":"Crane safety management system","version":"1.0.0"},"paths":{"/api/health/":{"get":{"tags":["health"],"summary":"Health Check Endpoint","description":"Health check endpoint for service monitoring.","operationId":"health_check_endpoint_api_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheckResponse"}}}}}}},"/api/health/reset-transactional":{"post":{"tags":["health"],"summary":"Reset Transactional Data Endpoint","description":"Reset only the transactional data in the database.\nPreserves master data like users, cranes, etc.","operationId":"reset_transactional_data_endpoint_api_health_reset_transactional_post","responses":{"204":{"description":"Successful Response"}}}},"/api/health/reset-full":{"post":{"tags":["health"],"summary":"Reset Full Database For Testing Endpoint","description":"Reset the entire database. FOR DEVELOPMENT/TESTING ONLY.\nWARNING: This is a destructive operation.","operationId":"reset_full_database_for_testing_endpoint_api_health_reset_full_post","responses":{"204":{"description":"Successful Response"}}}},"/api/sites":{"post":{"tags":["sites"],"summary":"Create Site Endpoint","description":"Create a new construction site.","operationId":"create_site_endpoint_api_sites_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SiteCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SiteOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["sites"],"summary":"List Sites Endpoint","description":"List all construction sites.\nIf 'mine' is true, returns only sites relevant to the user_id.\nNOTE: In a real app, user_id would come from an auth dependency.","operationId":"list_sites_endpoint_api_sites_get","parameters":[{"name":"mine","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Mine"}},{"name":"user_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SiteOut"},"title":"Response List Sites Endpoint Api Sites Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sites/{site_id}":{"patch":{"tags":["sites"],"summary":"Update Site Endpoint","description":"Update a site's attributes, including approving it.","operationId":"update_site_endpoint_api_sites__site_id__patch","parameters":[{"name":"site_id","in":"path","required":true,"schema":{"type":"string","title":"Site Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SiteUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SiteOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cranes":{"get":{"tags":["cranes"],"summary":"List Cranes Endpoint","description":"List all cranes with optional filtering.","operationId":"list_cranes_endpoint_api_cranes_get","parameters":[{"name":"owner_org_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Owner Org Id"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/CraneStatus"},{"type":"null"}],"title":"Status"}},{"name":"model_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Name"}},{"name":"min_capacity","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Min Capacity"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CraneOut"},"title":"Response List Cranes Endpoint Api Cranes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/crane-models":{"get":{"tags":["crane-models"],"summary":"List Crane Models Endpoint","description":"List all crane models.","operationId":"list_crane_models_endpoint_api_crane_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CraneModelOut"},"type":"array","title":"Response List Crane Models Endpoint Api Crane Models Get"}}}}}}},"/api/owners/with-stats":{"get":{"tags":["owners"],"summary":"List Owners With Stats Endpoint","description":"List all owners with statistics about their crane fleet.","operationId":"list_owners_with_stats_endpoint_api_owners_with_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OwnerStatsOut"},"type":"array","title":"Response List Owners With Stats Endpoint Api Owners With Stats Get"}}}}}}},"/api/owners/{owner_id}/cranes":{"get":{"tags":["owners"],"summary":"List Owner Cranes Endpoint","description":"List all cranes owned by a specific organization, with optional filters.","operationId":"list_owner_cranes_endpoint_api_owners__owner_id__cranes_get","parameters":[{"name":"owner_id","in":"path","required":true,"schema":{"type":"string","title":"Owner Id"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/CraneStatus"},{"type":"null"}],"title":"Status"}},{"name":"model_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Name"}},{"name":"min_capacity","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Min Capacity"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CraneOut"},"title":"Response List Owner Cranes Endpoint Api Owners  Owner Id  Cranes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/owners/me/requests":{"get":{"tags":["owners"],"summary":"List My Requests Endpoint","description":"List requests for the current owner.","operationId":"list_my_requests_endpoint_api_owners_me_requests_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/RequestType"},{"type":"null"}],"title":"Type"}},{"name":"request_status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/RequestStatus"},{"type":"null"}],"title":"Request Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RequestOut"},"title":"Response List My Requests Endpoint Api Owners Me Requests Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/crane-assignments":{"post":{"tags":["crane-assignments"],"summary":"Create Crane Assignment Endpoint","description":"Create a new crane assignment to a site.","operationId":"create_crane_assignment_endpoint_api_crane_assignments_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignCraneIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignmentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/driver-assignments":{"post":{"tags":["driver-assignments"],"summary":"Create Driver Assignment Endpoint","description":"Assign a driver to a site-crane combination.","operationId":"create_driver_assignment_endpoint_api_driver_assignments_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignDriverIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DriverAssignmentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/attendances":{"post":{"tags":["attendances"],"summary":"Create Attendance Endpoint","description":"Record a new attendance entry for a driver.","operationId":"create_attendance_endpoint_api_attendances_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/document-requests":{"post":{"tags":["document-requests"],"summary":"Create Document Request Endpoint","description":"Request a document from a driver.","operationId":"create_document_request_endpoint_api_document_requests_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocRequestIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocRequestResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/document-items":{"post":{"tags":["document-items"],"summary":"Submit Document Item Endpoint","description":"Submit a new document item.","operationId":"submit_document_item_endpoint_api_document_items_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocItemSubmitIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocSubmitResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/document-items/{item_id}":{"patch":{"tags":["document-items"],"summary":"Review Document Item Endpoint","description":"Review (approve or reject) a document item.","operationId":"review_document_item_endpoint_api_document_items__item_id__patch","parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocItemReviewIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocItemResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/requests":{"post":{"tags":["requests"],"summary":"Create Request Endpoint","description":"Create a new request, e.g. for crane deployment.","operationId":"create_request_endpoint_api_requests_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/requests/{request_id}/respond":{"post":{"tags":["requests"],"summary":"Respond To Request Endpoint","description":"Approve or reject a deployment request.\nThis is typically done by an OWNER.","operationId":"respond_to_request_endpoint_api_requests__request_id__respond_post","parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"string","title":"Request Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AssignCraneIn":{"properties":{"site_id":{"type":"string","title":"Site Id","description":"Target site ID"},"crane_id":{"type":"string","title":"Crane Id","description":"Crane to assign"},"safety_manager_id":{"type":"string","title":"Safety Manager Id","description":"SAFETY_MANAGER user ID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"Assignment start date"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"Assignment end date"}},"type":"object","required":["site_id","crane_id","safety_manager_id","start_date"],"title":"AssignCraneIn","description":"Schema for assigning a crane to a site."},"AssignDriverIn":{"properties":{"site_crane_id":{"type":"string","title":"Site Crane Id","description":"Site-crane assignment ID"},"driver_id":{"type":"string","title":"Driver Id","description":"DRIVER user ID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"Driver assignment start date"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"Driver assignment end date"}},"type":"object","required":["site_crane_id","driver_id","start_date"],"title":"AssignDriverIn","description":"Schema for assigning a driver to a site-crane combination."},"AssignmentResponse":{"properties":{"assignment_id":{"type":"string","title":"Assignment Id","description":"Created assignment ID"}},"type":"object","required":["assignment_id"],"title":"AssignmentResponse","description":"Schema for crane assignment operation responses."},"AttendanceIn":{"properties":{"driver_assignment_id":{"type":"string","title":"Driver Assignment Id","description":"Driver assignment ID"},"work_date":{"type":"string","format":"date","title":"Work Date","description":"Work date"},"check_in_at":{"type":"string","format":"date-time","title":"Check In At","description":"Check-in timestamp"},"check_out_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Check Out At","description":"Check-out timestamp"}},"type":"object","required":["driver_assignment_id","work_date","check_in_at"],"title":"AttendanceIn","description":"Schema for recording driver attendance."},"AttendanceResponse":{"properties":{"attendance_id":{"type":"string","title":"Attendance Id","description":"Created attendance record ID"}},"type":"object","required":["attendance_id"],"title":"AttendanceResponse","description":"Schema for attendance recording responses."},"CraneModelOut":{"properties":{"id":{"type":"string","title":"Id"},"model_name":{"type":"string","title":"Model Name"},"max_lifting_capacity_ton_m":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Lifting Capacity Ton M"},"max_working_height_m":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max Working Height M"},"max_working_radius_m":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max Working Radius M"},"optional_specs":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Optional Specs"}},"type":"object","required":["id","model_name"],"title":"CraneModelOut","description":"Schema for crane model specifications in API responses."},"CraneOut":{"properties":{"id":{"type":"string","title":"Id"},"owner_org_id":{"type":"string","title":"Owner Org Id"},"serial_no":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Serial No"},"status":{"$ref":"#/components/schemas/CraneStatus"},"model":{"$ref":"#/components/schemas/CraneModelOut"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","owner_org_id","status","model","created_at","updated_at"],"title":"CraneOut","description":"Schema for crane information in API responses."},"CraneStatus":{"type":"string","enum":["NORMAL","REPAIR","INBOUND"],"title":"CraneStatus","description":"Crane operational status values."},"DocItemResponse":{"properties":{"item_id":{"type":"string","title":"Item Id","description":"Document item ID"},"status":{"$ref":"#/components/schemas/DocItemStatus","description":"Current document status"}},"type":"object","required":["item_id","status"],"title":"DocItemResponse","description":"Schema for document review responses."},"DocItemReviewIn":{"properties":{"item_id":{"type":"string","title":"Item Id","description":"Document item ID"},"reviewer_id":{"type":"string","title":"Reviewer Id","description":"SAFETY_MANAGER user ID"},"approve":{"type":"boolean","title":"Approve","description":"True to approve, False to reject"}},"type":"object","required":["item_id","reviewer_id","approve"],"title":"DocItemReviewIn","description":"Schema for reviewing a document item."},"DocItemStatus":{"type":"string","enum":["PENDING","SUBMITTED","APPROVED","REJECTED"],"title":"DocItemStatus","description":"Document item review status."},"DocItemSubmitIn":{"properties":{"request_id":{"type":"string","title":"Request Id","description":"Document request ID"},"doc_type":{"type":"string","minLength":1,"title":"Doc Type","description":"Document type"},"file_url":{"type":"string","minLength":1,"format":"uri","title":"File Url","description":"HTTPS URL to document file"}},"type":"object","required":["request_id","doc_type","file_url"],"title":"DocItemSubmitIn","description":"Schema for submitting a document item."},"DocRequestIn":{"properties":{"site_id":{"type":"string","title":"Site Id","description":"Site ID"},"driver_id":{"type":"string","title":"Driver Id","description":"Driver user ID"},"requested_by_id":{"type":"string","title":"Requested By Id","description":"SAFETY_MANAGER user ID"},"due_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Due Date","description":"Document submission due date"}},"type":"object","required":["site_id","driver_id","requested_by_id"],"title":"DocRequestIn","description":"Schema for creating a document request."},"DocRequestResponse":{"properties":{"request_id":{"type":"string","title":"Request Id","description":"Created document request ID"}},"type":"object","required":["request_id"],"title":"DocRequestResponse","description":"Schema for document request creation responses."},"DocSubmitResponse":{"properties":{"item_id":{"type":"string","title":"Item Id","description":"Created document item ID"}},"type":"object","required":["item_id"],"title":"DocSubmitResponse","description":"Schema for document submission responses."},"DriverAssignmentResponse":{"properties":{"driver_assignment_id":{"type":"string","title":"Driver Assignment Id","description":"Created driver assignment ID"}},"type":"object","required":["driver_assignment_id"],"title":"DriverAssignmentResponse","description":"Schema for driver assignment operation responses."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthCheckResponse":{"properties":{"status":{"type":"string","title":"Status","description":"Service status"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp","description":"Current server time"},"database_healthy":{"type":"boolean","title":"Database Healthy","description":"Database connectivity status"}},"type":"object","required":["status","timestamp","database_healthy"],"title":"HealthCheckResponse","description":"Schema for health check endpoint response."},"OwnerStatsOut":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"total_cranes":{"type":"integer","title":"Total Cranes"},"available_cranes":{"type":"integer","title":"Available Cranes"}},"type":"object","required":["id","name","total_cranes","available_cranes"],"title":"OwnerStatsOut","description":"Schema for an owner organization with crane fleet statistics."},"RequestCreate":{"properties":{"type":{"$ref":"#/components/schemas/RequestType"},"requester_id":{"type":"string","title":"Requester Id"},"target_entity_id":{"type":"string","title":"Target Entity Id"},"related_entity_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Related Entity Id"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["type","requester_id","target_entity_id"],"title":"RequestCreate","description":"Schema for creating a generic request."},"RequestOut":{"properties":{"id":{"type":"string","title":"Id"},"type":{"$ref":"#/components/schemas/RequestType"},"status":{"$ref":"#/components/schemas/RequestStatus"},"requester_id":{"type":"string","title":"Requester Id"},"approver_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Approver Id"},"target_entity_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Entity Id"},"related_entity_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Related Entity Id"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"requested_at":{"type":"string","format":"date-time","title":"Requested At"},"responded_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Responded At"}},"type":"object","required":["id","type","status","requester_id","requested_at"],"title":"RequestOut","description":"Schema for request information in API responses."},"RequestStatus":{"type":"string","enum":["PENDING","APPROVED","REJECTED"],"title":"RequestStatus","description":"Generic request statuses."},"RequestType":{"type":"string","enum":["CRANE_DEPLOY"],"title":"RequestType","description":"Generic request types."},"RequestUpdate":{"properties":{"status":{"$ref":"#/components/schemas/RequestStatus"},"approver_id":{"type":"string","title":"Approver Id"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["status","approver_id"],"title":"RequestUpdate","description":"Schema for approving or rejecting a request."},"SiteCreate":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name","description":"Site name"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address","description":"Site address"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"Project start date"},"end_date":{"type":"string","format":"date","title":"End Date","description":"Project end date"},"requested_by_id":{"type":"string","title":"Requested By Id","description":"SAFETY_MANAGER user ID"}},"type":"object","required":["name","start_date","end_date","requested_by_id"],"title":"SiteCreate","description":"Schema for creating a new construction site.","example":{"address":"123 Main Street, Seoul","end_date":"2025-12-31","name":"Downtown Construction Site","requested_by_id":"safety-manager-uuid","start_date":"2025-09-01"}},"SiteOut":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"status":{"$ref":"#/components/schemas/SiteStatus"},"requested_by_id":{"type":"string","title":"Requested By Id"},"approved_by_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Approved By Id"},"requested_at":{"type":"string","format":"date-time","title":"Requested At"},"approved_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Approved At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","name","start_date","end_date","status","requested_by_id","requested_at","created_at","updated_at"],"title":"SiteOut","description":"Schema for site information in API responses."},"SiteStatus":{"type":"string","enum":["PENDING_APPROVAL","ACTIVE","REJECTED","COMPLETED"],"title":"SiteStatus","description":"Construction site status values."},"SiteUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"start_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Start Date"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date"},"status":{"anyOf":[{"$ref":"#/components/schemas/SiteStatus"},{"type":"null"}]},"approved_by_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Approved By Id"},"approved_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Approved At"}},"type":"object","title":"SiteUpdate","description":"Schema for updating a site. All fields are optional."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}